+++ src/hooks/player_object.cpp
@@ -63,3 +63,20 @@ void HookedPlayerObject::playSpiderDashEffect(cocos2d::CCPoint from, cocos2d::CC
 
     PlayerObject::playSpiderDashEffect(from, to);
 }
+
+void HookedPlayerObject::incrementJumps() {
+    if (PlayLayer::get() == nullptr) {
+        PlayerObject::incrementJumps();
+        return;
+    }
+
+    auto* gpl = static_cast<GlobedPlayLayer*>(PlayLayer::get());
+
+    if (this == gpl->m_player1) {
+        gpl->m_fields->didJustJumpp1 = true;
+    } else if (this == gpl->m_player2) {
+        gpl->m_fields->didJustJumpp2 = true;
+    }
+
+    PlayerObject::incrementJumps();
+}