+++ src/ui/game/player/base_visual_player.hpp
@@ -20,6 +20,7 @@ public:
     virtual void updateIconType(PlayerIconType newType) = 0;
     virtual void playDeathEffect() = 0;
     virtual void playSpiderTeleport(const SpiderTeleportData& data) = 0;
+    virtual void playJump() = 0;
     virtual cocos2d::CCPoint getPlayerPosition() = 0;
 
     static int getIconWithType(const PlayerIconData& data, PlayerIconType type);